+++ datafusion/core/tests/fuzz_cases/window_fuzz.rs
@@ -30,6 +30,7 @@ use datafusion::physical_plan::windows::{
 use datafusion::physical_plan::{collect, ExecutionPlan, InputOrderMode};
 use datafusion::prelude::{SessionConfig, SessionContext};
 use datafusion_common::{Result, ScalarValue};
+use datafusion_common_runtime::SpawnedTask;
 use datafusion_expr::type_coercion::aggregates::coerce_types;
 use datafusion_expr::{
     AggregateFunction, BuiltInWindowFunction, WindowFrame, WindowFrameBound,
@@ -123,8 +124,7 @@ async fn window_bounded_window_random_comparison() -> Result<()> {
         for i in 0..n {
             let idx = i % test_cases.len();
             let (pb_cols, ob_cols, search_mode) = test_cases[idx].clone();
-            #[allow(clippy::disallowed_methods)] // spawn allowed only in tests
-            let job = tokio::spawn(run_window_test(
+            let job = SpawnedTask::spawn(run_window_test(
                 make_staggered_batches::<true>(1000, n_distinct, i as u64),
                 i as u64,
                 pb_cols,
@@ -134,7 +134,7 @@ async fn window_bounded_window_random_comparison() -> Result<()> {
             handles.push(job);
         }
         for job in handles {
-            job.await.unwrap()?;
+            job.join().await.unwrap()?;
         }
     }
     Ok(())
