+++ datafusion/physical-plan/src/repartition/mod.rs
@@ -29,7 +29,7 @@ use super::metrics::{self, ExecutionPlanMetricsSet, MetricBuilder, MetricsSet};
 use super::{
     DisplayAs, ExecutionPlanProperties, RecordBatchStream, SendableRecordBatchStream,
 };
-use crate::common::{transpose, SpawnedTask};
+use crate::common::transpose;
 use crate::hash_utils::create_hashes;
 use crate::metrics::BaselineMetrics;
 use crate::repartition::distributor_channels::{
@@ -42,6 +42,7 @@ use arrow::array::{ArrayRef, UInt64Builder};
 use arrow::datatypes::SchemaRef;
 use arrow::record_batch::RecordBatch;
 use datafusion_common::{arrow_datafusion_err, not_impl_err, DataFusionError, Result};
+use datafusion_common_runtime::SpawnedTask;
 use datafusion_execution::memory_pool::MemoryConsumer;
 use datafusion_execution::TaskContext;
 use datafusion_physical_expr::{EquivalenceProperties, PhysicalExpr, PhysicalSortExpr};
@@ -946,7 +947,6 @@ mod tests {
     use datafusion_execution::runtime_env::{RuntimeConfig, RuntimeEnv};
 
     use futures::FutureExt;
-    use tokio::task::JoinHandle;
 
     #[tokio::test]
     async fn one_to_many_round_robin() -> Result<()> {
@@ -1060,10 +1060,9 @@ mod tests {
     }
 
     #[tokio::test]
-    #[allow(clippy::disallowed_methods)]
     async fn many_to_many_round_robin_within_tokio_task() -> Result<()> {
-        let join_handle: JoinHandle<Result<Vec<Vec<RecordBatch>>>> =
-            tokio::spawn(async move {
+        let handle: SpawnedTask<Result<Vec<Vec<RecordBatch>>>> =
+            SpawnedTask::spawn(async move {
                 // define input partitions
                 let schema = test_schema();
                 let partition = create_vec_batches(50);
@@ -1074,7 +1073,7 @@ mod tests {
                 repartition(&schema, partitions, Partitioning::RoundRobinBatch(5)).await
             });
 
-        let output_partitions = join_handle.await.unwrap().unwrap();
+        let output_partitions = handle.join().await.unwrap().unwrap();
 
         assert_eq!(5, output_partitions.len());
         assert_eq!(30, output_partitions[0].len());
