+++ datafusion/sqllogictest/Cargo.toml
@@ -39,6 +39,7 @@ chrono = { workspace = true, optional = true }
 clap = { version = "4.4.8", features = ["derive", "env"] }
 datafusion = { workspace = true, default-features = true }
 datafusion-common = { workspace = true, default-features = true }
+datafusion-common-runtime = { workspace = true, default-features = true }
 futures = { workspace = true }
 half = { workspace = true, default-features = true }
 itertools = { workspace = true }
